requirements.txt
streamlit
torch
torchvision
matplotlib
opencv-python-headless
requests
Pillow
